requirements.txt
beautifulsoup4==4.13.3
requests==2.32.3
python-dotenv==1.0.1
